feat(web-video-action): add user action queries
Align video action type values with short database fields.

Add mapper queries for user action counts and records while reusing the shared login-state helper.

+++ nilo-common/src/main/java/com/neon/nilocommon/entity/enums/userVideoAction/VideoActionType.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum VideoActionType
 {
-    LIKE(1), SAVE(2), COIN(3);
+    LIKE((short) 1), SAVE((short) 2), COIN((short) 3);
 
-    private final int value;
+    private final short value;
 }
