+++ nilo-common/src/main/java/com/neon/nilocommon/entity/query/UserVideoActionQuery.java
@@ -32,7 +32,7 @@ public class UserVideoActionQuery extends BaseQuery
     /**
      * 1:视频点赞 2:视频收藏 3:视频投币
      */
-    private Integer actionType;
+    private Short actionType;
 
     /**
      * 投币数量
@@ -52,6 +52,4 @@ public class UserVideoActionQuery extends BaseQuery
     private String actionTimeStart;
 
     private String actionTimeEnd;
-
-
 }
