+++ nilo-common/src/main/java/com/neon/nilocommon/entity/vo/UserVideoActionVO.java
@@ -12,7 +12,7 @@ public class UserVideoActionVO
     /**
      * 1:视频点赞 2:视频收藏 3:视频投币
      */
-    private Integer actionType;
+    private Short actionType;
 
     /**
      * 投币数量
